feat: refactor into separate component
